listReferenceBases function

Get the sequence bases of a reference genome by genomic range.

listReferenceBases(host, referenceId, start = 1, end = NA_integer_,
  pageToken = NA_character_)

Arguments

host

URL of GA4GH API data server.

referenceId

The ID of the Reference to be retrieved.

start

The start position (1-based) of this query. Defaults to 0. Genomic positions are non-negative integers less than reference length. Requests spanning the join of circular genomes are represented as two requests one on each side of the join (position 1).

end

The end position (1-based, exclusive) of this query. Defaults to the length of this Reference.

pageToken

Used to move through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.

Value

List of genomic sequences.

Details

This function requests POST host/listreferencebases.

Examples

host <- "http://1kgenomes.ga4gh.org/"
## Not run: ------------------------------------
# referenceSetId <- searchReferenceSets(host, pageSize = 1)$id
# referenceId <- searchReferences(host, referenceSetId, pageSize = 1)$id
# listReferenceBases(host, referenceId, start = 1, end = 100)
## ---------------------------------------------